Issue #12: Fix a bug causing the application to crash when the remote name for the local repository is 'local' instead of being ''.
This is actually an anomaly, and we should standardize how the local remote is
represented. Unfortunately, we don't do this currently. Sometimes the local
repository is named 'local' and sometimes it's named ''. Probably explicitly
calling out that it's a local repo would be better.

In any case, this patch just fixes the symptoms, so it's suboptimal.